+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Dependencies
 - (backend) upgraded Express 4 to Express 5 ([#41](https://github.com/isolinear-labs/Neosynth/pull/41))
 - (backend/server) updated catch-all route to Express 5 wildcard syntax `/{*path}` ([#41](https://github.com/isolinear-labs/Neosynth/pull/41))
+- (backend) upgraded express-rate-limit from  7.5.0 to 8.3.1 ([#43](https://github.com/isolinear-labs/Neosynth/pull/43))
+- (backend) migrated express-rate-limit `max` option to `limit` for v8 compatibility ([#43](https://github.com/isolinear-labs/Neosynth/pull/43))
 
 ## [v1.0.3a] - 2025-11-21