+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Enhancements
 
 ### Bug Fixes
+
+### Dependencies
+
+## [v1.0.4] - 2026-03-28
+
+### Bug Fixes
 - (frontend/playlist) fixed playlistId not being set when saving playlists, preventing resume from loading full playlist context ([#30](https://github.com/isolinear-labs/Neosynth/pull/30))
 - (backend/models) removed duplicate schema indexes on fields already declared with `unique: true` ([#46](https://github.com/isolinear-labs/Neosynth/pull/46))